feat(categories): full CRUD with slug generation and unique constraints
- Add categories repo/service/controller; public GET, admin-only write.
- Joi validations; slugify util; map ER_DUP_ENTRY to 409.
- Routes refactor to per-path .route().all(methodNotAllowed).
- OpenAPI updated; Postman test steps documented.

+++ migrations/005_create_comments.js
@@ -1,7 +1,7 @@
 export async function up(knex) {
   await knex.schema.createTable('comments', t => {
-    t.bigIncrements('id').primary()
-    t.bigInteger('post_id')
+    t.increments('id').primary()
+    t.integer('post_id')
       .unsigned()
       .notNullable()
       .references('id')
